Repo: Added comparison operators and hash operator including test Cmd: AutoInterrupt handles boundary cases more gracefully as it can be that the os module suddenly becomes None if the interpreter is going down

+++ lib/git/cmd.py
@@ -63,8 +63,12 @@ class Git(object):
         def __del__(self):
             # did the process finish already so we have a return code ?
             if self.proc.poll() is not None:
-                return 
-            
+                return
+                
+            # can be that nothing really exists anymore ... 
+            if os is None:
+                return
+                
             # try to kill it
             try:
                 os.kill(self.proc.pid, 2)   # interrupt signal
